doc-source/requirements.txt
attr-utils>=0.8.0
default-values>=0.5.0
extras-require>=0.2.0
furo>=2020.11.19b18
html-section>=0.1.0
seed-intersphinx-mapping>=0.3.1
sphinx>=3.0.3
sphinx-copybutton>=0.2.12
sphinx-debuginfo>=0.1.0
sphinx-licenseinfo>=0.1.1
sphinx-notfound-page>=0.5
sphinx-prompt>=1.1.0
sphinx-pyproject>=0.1.0
sphinx-tabs>=1.1.13
sphinx-toolbox>=2.13.0
sphinxemoji>=0.1.6
toctree-plus>=0.5.0
